Split crate into modules
